src/apicalypso/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
gunicorn==25.3.0
pydantic==2.13.3
sqlmodel==0.0.24
SQLAlchemy==2.0.49
asyncpg==0.31.0
contextlib2==21.6.0
slowapi==0.1.9
cryptography==47.0.0
bcrypt==5.0.0
python-jose[cryptography]
psycopg2-binary==2.9.12
python-multipart==0.0.27
httpx==0.28.1
requests==2.33.1
pytz==2026.1.post1
email-validator==2.3.0
pandas==3.0.2
pyarrow==24.0.0
orjson==3.11.8